playerListのデータをページに移動
