+++ src/types/PlayerSetting.d.ts
@@ -1,0 +1,4 @@
+type PlayerSetting = {
+  name: string;
+  colorHue: number;
+};
